refactor(serializers): move column mapping profile serializer
